refactor(actor-core): separate delivery internal impl and receptionist runtime
