+++ modules/actor-core/src/core/typed/delivery/internal.rs
@@ -1,0 +1,7 @@
+//! Internal implementation details for typed delivery controllers.
+
+mod producer_controller;
+mod work_pulling_producer_controller;
+
+pub(crate) use producer_controller::ProducerController;
+pub(crate) use work_pulling_producer_controller::WorkPullingProducerController;
